week1/1_instruction_generator/requirements.txt
python-dotenv==1.1.1
pandas==2.3.2
openai==1.101.0